+++ example/android/app/src/main/kotlin/com/example/flutter_awesome_context_menu_examples/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.flutter_awesome_context_menu_examples
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()