+++ lib/src/awesome_platform_utils.dart
@@ -2,7 +2,8 @@ import 'package:flutter/foundation.dart' show kIsWeb;
 import 'awesome_context_menu_item.dart';
 
 // Use conditional imports to handle platform-specific code
-import 'platform/platform_io.dart' if (dart.library.js) 'platform/platform_web.dart';
+import 'platform/platform_io.dart'
+    if (dart.library.js) 'platform/platform_web.dart';
 
 /// Utility class to handle platform-specific logic and operations
 class AwesomePlatformUtils {
@@ -13,8 +14,9 @@ class AwesomePlatformUtils {
   }
 
   /// Returns whether the app is running on a mobile platform (iOS or Android)
+  /// Also returns true for mobile browsers on web
   static bool isMobile() {
-    if (kIsWeb) return false;
+    // Check mobile platforms including mobile browsers
     return isMobilePlatform();
   }
 
@@ -37,6 +39,8 @@ class AwesomePlatformUtils {
 
   /// Get the appropriate submenu interaction mode for the current platform.
   static SubMenuInteractionMode getPlatformDefaultInteractionMode() {
-    return isHoverPlatform ? SubMenuInteractionMode.hover : SubMenuInteractionMode.click;
+    return isHoverPlatform
+        ? SubMenuInteractionMode.hover
+        : SubMenuInteractionMode.click;
   }
 }